+++ src/azure-cli/HISTORY.rst
@@ -35,9 +35,11 @@ Release History
 * vm create: Add --enable-agent configuration.
 * vm create: Use standard public IP SKU automatically when using zones.
 * vm create: Compose a valid computer name from VM name if computer name is not provided.
+* vm create: Add --workspace to enable log analytics workspace automatically.
+* vm extension set: Move --name and --extension-instance-name out of Resource Id Arguments group.
 * vmss create: Add --computer-name-prefix parameter to support custom computer name prefix of virtual machines in the VMSS.
 * Update galleries API version to 2019-07-01.
-* vm create: Add --workspace to enable log analytics workspace automatically.
+
 
 **Network**